api/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.0
sentry-sdk[fastapi]==1.45.0
python-dotenv==1.0.1
